+++ packages/form-input-serialize/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "react-layout-form",
+  "name": "form-input-serialize",
   "version": "1.0.0",
-  "description": "A form layout generator for your React components",
+  "description": "Serialize/Deserialize form input, and lookup",
   "main": "build/index.js",
   "author": "code@alvinsj.net",
   "repository": {
     "type": "git",
-    "url": "https://github.com/alvinsj/react-layout-form"
+    "url": "https://github.com/alvinsj/react-layout-builder"
   },
   "keywords": [
     "reactjs",
@@ -17,13 +17,12 @@
   ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/alvinsj/react-layout-form/issues"
+    "url": "https://github.com/alvinsj/react-layout-builder/issues"
   },
-  "homepage": "https://github.com/alvinsj/react-layout-form",
+  "homepage": "https://github.com/alvinsj/react-layout-builder",
   "dependencies": {
-    "classnames": "^2.1.3",
-    "object.assign": "^4.0.1",
-    "react-layout-builder": "^1.x.x"
+    "object-assign": "^4.x.x",
+    "form-serialize": "alvinsj/form-serialize#export-fn"
   },
   "devDependencies": {
     "babel": "^6.5.x",
@@ -39,12 +38,10 @@
     "react-addons-test-utils": "^15.3.2",
     "react-dom": "^15.3.2"
   },
-  "peerDependencies": {
-    "react": "^15.0.0"
-  },
   "scripts": {
     "jest": "NODE_PATH=src:NODE_PATH jest",
     "test": "jest",
+    "test:watch": "npm run jest -- --watch",
     "compile": "babel -d build/ src/",
     "prepublish": "npm run compile"
   },
