+++ packages/react-layout-build-form/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["es2015", "react"]
+}
+
