+++ packages/react-layout-build-form/__tests__/.eslintrc
@@ -1,0 +1,31 @@
+{
+    "extends": "eslint:recommended",
+    "rules": {
+        "indent": 2,
+        "no-unused-vars": 1,
+        "quotes": [
+            2,
+            "single"
+        ],
+        "semi": [
+            2,
+            "always"
+        ],
+        "no-mixed-spaces-and-tabs": 2,
+        "react/no-find-dom-node": 0
+    },
+    "globals": {
+        "jasmine": true,
+        "describe": true,
+        "xdescribe": true,
+        "jest": true,
+        "it": true,
+        "xit": true,
+        "expect": true,
+        "beforeEach": true,
+        "global": true,
+        "pending": true,
+        "setImmediate": true,
+        "afterEach": true
+    }
+}
