+++ packages/react-layout-build-form/package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-layout-builder",
   "version": "1.0.0",
-  "description": "Build form layout",
+  "description": "A helper to build layout",
   "main": "build/index.js",
   "author": "code@alvinsj.net",
   "repository": {
@@ -20,11 +20,10 @@
     "url": "https://github.com/alvinsj/react-layout-builder/issues"
   },
   "homepage": "https://github.com/alvinsj/react-layout-builder",
-  "peerDependencies": {
-    "react": "^15.x.x"
-  },
   "dependencies": {
-    "object-assign": "^4.x.x",
+    "classnames": "^2.1.3",
+    "object-assign": "^4.0.1",
+    "react-layout-builder": "^1.x.x",
     "form-input-serialize": "^1.x.x"
   },
   "devDependencies": {
@@ -41,10 +40,12 @@
     "react-addons-test-utils": "^15.3.2",
     "react-dom": "^15.3.2"
   },
+  "peerDependencies": {
+    "react": "^15.0.0"
+  },
   "scripts": {
     "jest": "NODE_PATH=src:NODE_PATH jest",
     "test": "jest",
-    "test:watch": "npm run jest -- --watch",
     "compile": "babel -d build/ src/",
     "prepublish": "npm run compile"
   },
