+++ packages/react-layout-builder/src/index.js
@@ -1,14 +1,12 @@
 import React from 'react';
-import assign from 'object.assign';
-import serialize, {hash_serializer} from 'form-serialize';
 
 var layoutCount = 0;
-const renderLayout = (...children) => {
+export const layout = (...children) => {
     return <div key={`layout-${layoutCount++}`} className="layout">{children}</div>;
 };
 
 var sectionCount = 0;
-const renderSection = (name, ...rows) => {
+export const section = (name, ...rows) => {
     const renderedRows = rows.map((cols, index) => {
         return <div key={`columns-${index}`} className="columns">{cols}</div>;
     });
@@ -19,7 +17,7 @@ const renderSection = (name, ...rows) => {
         </div>);
 };
 
-const renderCol = (renderField, type, ...fields) => {
+export const col = (renderField, type, ...fields) => {
     if(fields.length == 0)
         return <div className={`column ${type}`} />;
 
@@ -32,69 +30,3 @@ const renderCol = (renderField, type, ...fields) => {
         </div>);
     });
 };
-
-export const builder = {
-    layout: renderLayout,
-    section: renderSection,
-    col: renderCol
-};
-
-export function inputPropsLookup (inputProps, inputName){
-    var props = inputProps[inputName];
-    if(props) return assign({label: inputName}, props); // props found, easy.
-
-    var inputNameTree = hash_serializer({}, inputName, null);
-    var attrName = Object.keys(inputNameTree)[0]; // inputName describes only single path
-
-    props = inputProps[attrName];
-
-    if( !isNaN(parseInt(attrName)) ){ // skip number
-        props = inputProps;
-    }else if(props && props.type == 'nested'){ // continue with nested props
-        props = props.fields;
-    }else if(props){
-        return props;
-    }else throw (inputName+': props cannot be found.');
-
-    var newInputName = inputName.replace(attrName, '');
-    var nextAttrName =  Object.keys(inputNameTree[attrName])[0];
-
-    // remove bracket
-    newInputName = newInputName.replace('['+nextAttrName+']', '');
-    newInputName = nextAttrName + newInputName;
-
-    return inputPropsLookup(props, newInputName);
-}
-
-export function inputValueLookup(serializedValues, inputName){
-    if(!serializedValues || Object.keys(serializedValues).length == 0) return null;
-
-    var value = serializedValues[inputName];
-    if(value) return value; // value found, easy.
-
-    // try to serialize and traverse
-    var inputNameTree = hash_serializer({}, inputName, null);
-    var attrName = Object.keys(inputNameTree)[0]; // inputName describes only single path
-
-    var next = inputNameTree[attrName];
-    if(next == null) return null; // if it's already leaf, then no chance to be found.
-
-    var nextAttrName =  Object.keys(next)[0];
-    var newInputName = inputName.replace(attrName, '');
-
-    // remove bracket
-    newInputName = newInputName.replace('['+nextAttrName+']', '');
-
-    if(newInputName === ''){
-        return serializedValues[attrName] ?
-            serializedValues[attrName][nextAttrName] : undefined;
-    }
-    newInputName = nextAttrName + newInputName;
-
-    var nestedValues = serializedValues[attrName];
-    return inputValueLookup(nestedValues, newInputName);
-}
-
-export function formInputsSerialize(form){
-    return serialize(form, {hash: true});
-}
